requirements.txt
pandas
flask
flask-socketio
python-dotenv
kiteconnect
numpy
scikit-learn
xgboost
joblib
pyarrow
lightgbm
catboost
optuna
torch
stable-baselines3
psutil
yfinance
requests
beautifulsoup4
schedule
psycopg2-binary
hmmlearn
